openai-img/requirements.txt
openai>=1.14.0
requests>=2.31.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0 